+++ WizBot.Core/Modules/Administration/Services/UserPunishService.cs
@@ -215,11 +215,11 @@ WHERE GuildId={guildId}
             }
         }
 
-        public async Task<int> GetWarnExpire(ulong guildId)
+        public Task<int> GetWarnExpire(ulong guildId)
         {
             using var uow = _db.GetDbContext();
             var config = uow.GuildConfigs.ForId(guildId);
-            return config.WarnExpireHours / 24;
+            return Task.FromResult(config.WarnExpireHours / 24);
         }
 
         public async Task WarnExpireAsync(ulong guildId, int days, bool delete)