+++ WizBot.Core/Modules/Xp/Xp.cs
@@ -48,7 +48,7 @@ namespace WizBot.Modules.Xp
             if (!reply)
                 return;
 
-            _service.ResetXpRewards(ctx.Guild.Id);
+            await _service.ResetXpRewards(ctx.Guild.Id);
             await ctx.OkAsync();
         }